+++ verticapy/machine_learning/vertica/decomposition.py
@@ -20,7 +20,7 @@ permissions and limitations under the License.
 # Modules
 #
 # VerticaPy Modules
-from verticapy.decorators import (
+from verticapy.utils._decorators import (
     save_verticapy_logs,
     check_minimum_version,
 )
