+++ verticapy/machine_learning/vertica/delphi.py
@@ -25,7 +25,7 @@ import numpy as np
 from typing import Union, Literal
 
 # VerticaPy Modules
-from verticapy.decorators import (
+from verticapy.utils._decorators import (
     save_verticapy_logs,
     check_minimum_version,
 )
