+++ verticapy/machine_learning/vertica/ensemble.py
@@ -25,7 +25,7 @@ import random
 import numpy as np
 
 # VerticaPy Modules
-from verticapy.decorators import (
+from verticapy.utils._decorators import (
     save_verticapy_logs,
     check_minimum_version,
 )
