+++ verticapy/machine_learning/vertica/model_selection.py
@@ -28,7 +28,7 @@ from typing import Union, Literal
 
 # VerticaPy Modules
 import verticapy
-from verticapy.decorators import (
+from verticapy.utils._decorators import (
     save_verticapy_logs,
     check_minimum_version,
 )
