+++ verticapy/machine_learning/vertica/vmodel.py
@@ -26,7 +26,7 @@ from typing import Union, Literal
 
 # VerticaPy Modules
 import verticapy
-from verticapy.decorators import (
+from verticapy.utils._decorators import (
     save_verticapy_logs,
     check_minimum_version,
 )
