+++ verticapy/plotting/_matplotlib/mlplot.py
@@ -30,7 +30,7 @@ from matplotlib.lines import Line2D
 import numpy as np
 
 # VerticaPy Modules
-from verticapy.decorators import (
+from verticapy.utils._decorators import (
     save_verticapy_logs,
     check_minimum_version,
 )
