+++ verticapy/udf.py
@@ -25,7 +25,7 @@ from typing import Union, Literal
 
 # VerticaPy Modules
 import verticapy as vp
-from verticapy.decorators import (
+from verticapy.utils._decorators import (
     save_verticapy_logs,
     check_minimum_version,
 )
